Move llvm.rs and vs.rs to platform/windows.rs
