+++ skia-bindings/build_support/skia.rs
@@ -1,6 +1,4 @@
 mod config;
 pub mod env;
-pub mod llvm;
-pub mod vs;
 
 pub use config::{build, BuildConfiguration, FinalBuildConfiguration};
